feat(backend-tls): use tls between traefik and notary
This is a small consolidation of notary, we've moved the signer into the
same pod as the server and bind it to localhost so only processes within
the pod can communicate with the signer. Also we've now made sure that
the path from traefik to the notary server is encrypted with tls.